+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntityComponents.java
@@ -216,7 +216,8 @@ public final class EntityComponents {
 																									 Supplier<EntityEditPanel> editPanel) {
 		ForeignKeyDefinition foreignKeyDefinition = entityDefinition.foreignKeys().definition(foreignKey);
 
-		return EntityComboBoxPanel.builder(comboBoxModel)
+		return EntityComboBoxPanel.builder()
+						.model(comboBoxModel)
 						.editPanel(editPanel)
 						.toolTipText(foreignKeyDefinition.description().orElse(null));
 	}